fix(media): break media <-> regex circular import
`src/regex.ts` imported `Media` from `src/media.ts` and dereferenced
`Media.AvailableOutputFormats` at module top level, while `src/media.ts`
imported `contentTypeIsImageForSharp` from `src/regex.ts`.

Entering through the package barrel, `regex.js` finished evaluating first
and the cycle was harmless. Entering through the `/media` subpath directly,
`regex.js` was evaluated while `media.js` was still initialising, so `Media`
was still uninitialised and the import threw:

    TypeError: Cannot read properties of undefined (reading 'AvailableOutputFormats')

Extract `AvailableOutputFormats` into a leaf module (`src/media-formats.ts`)
that imports nothing from `media.ts` or `regex.ts`, and have both modules
import it from there. `Media.AvailableOutputFormats` is re-exported so the
public API, member order, and values are unchanged, and the generated
`contentTypeIsImageForSharp` pattern is byte-identical.

The cycle predates the `sharp` 0.35 bump; it is present at 55c450c.
