+++ lib/media.js
@@ -6,6 +6,7 @@ import { Buffer } from 'buffer';
 import { getStorage } from 'firebase-admin/storage';
 import { logger } from 'firebase-functions/v2';
 import sharp from 'sharp';
+import { AvailableOutputFormats as AvailableOutputFormatsLeaf } from './media-formats.js';
 import { contentTypeIsImageForSharp } from './regex.js';
 import { emulator } from './variables.js';
 /**
@@ -19,35 +20,13 @@ export var Media;
     /**
      * Available Image Output Formats
      *
-     * Members mirror the container formats libvips can read.  Only a subset is
-     * usable as an output format by `sharp.toFormat` (`avif`, `dz`, `gif`, `jpeg`,
-     * `jp2`, `jxl`, `png`, `raw`, `tiff`, `webp`); the remaining members are kept
-     * for backwards compatibility and for content-type detection, and sharp will
-     * reject them if requested as an output format.
+     * Re-exported from `./media-formats.js` so that `Media.AvailableOutputFormats`
+     * keeps resolving to exactly the same enum, with the same members and values,
+     * while the underlying declaration lives in a leaf module that does not
+     * participate in the `media` <-> `regex` import cycle.
      * @enum {string}
      */
-    let AvailableOutputFormats;
-    (function (AvailableOutputFormats) {
-        AvailableOutputFormats["avif"] = "avif";
-        AvailableOutputFormats["dz"] = "dz";
-        AvailableOutputFormats["fits"] = "fits";
-        AvailableOutputFormats["gif"] = "gif";
-        AvailableOutputFormats["heif"] = "heif";
-        AvailableOutputFormats["input"] = "input";
-        AvailableOutputFormats["jpeg"] = "jpeg";
-        AvailableOutputFormats["jp2"] = "jp2";
-        AvailableOutputFormats["jxl"] = "jxl";
-        AvailableOutputFormats["magick"] = "magick";
-        AvailableOutputFormats["openslide"] = "openslide";
-        AvailableOutputFormats["pdf"] = "pdf";
-        AvailableOutputFormats["png"] = "png";
-        AvailableOutputFormats["ppm"] = "ppm";
-        AvailableOutputFormats["raw"] = "raw";
-        AvailableOutputFormats["svg"] = "svg";
-        AvailableOutputFormats["tiff"] = "tiff";
-        AvailableOutputFormats["v"] = "v";
-        AvailableOutputFormats["webp"] = "webp";
-    })(AvailableOutputFormats = Media.AvailableOutputFormats || (Media.AvailableOutputFormats = {}));
+    Media.AvailableOutputFormats = AvailableOutputFormatsLeaf;
     /**
      * Set of predefined image sizes
      * @enum {string}
@@ -335,7 +314,7 @@ export var Media;
             // Show default image if media file is not found
             const media = await Image.resize({
                 fileName: 'default/error.jpg',
-                format: AvailableOutputFormats.jpeg,
+                format: Media.AvailableOutputFormats.jpeg,
                 ...imageResizeOptions,
             });
             mediaBuffer = media.buffer;
@@ -403,7 +382,7 @@ export var Media;
     Image.bufferImage = async (options) => {
         let outputFormat;
         const convertToFormatsEnum = (str) => {
-            const colorValue = AvailableOutputFormats[str];
+            const colorValue = Media.AvailableOutputFormats[str];
             return colorValue;
         };
         if (options.contentType) {
@@ -413,8 +392,8 @@ export var Media;
         if (options.format) {
             outputFormat = convertToFormatsEnum(options.format);
         }
-        outputFormat ?? (outputFormat = AvailableOutputFormats.jpeg);
-        const animated = outputFormat === AvailableOutputFormats.gif;
+        outputFormat ?? (outputFormat = Media.AvailableOutputFormats.jpeg);
+        const animated = outputFormat === Media.AvailableOutputFormats.gif;
         let optionsImage = {};
         let dpr = options.dpr ?? 1;
         if (dpr > 6)
