+++ lib/regex.js
@@ -2,7 +2,7 @@
  * @license
  * Copyright FabricElements. All Rights Reserved.
  */
-import { Media } from './media.js';
+import { AvailableOutputFormats } from './media-formats.js';
 /* eslint-disable max-len, no-control-regex */
 /**
  * Regular expression that matches a valid RFC 5322-compliant email address.
@@ -15,7 +15,7 @@ export const isEmail = new RegExp('(?:[a-z0-9!#$%&\'*+/=?^_`{|}~-]+(?:\\.[a-z0-9
  * an image format that the `sharp` library can process.
  * The accepted format list is derived dynamically from `Media.AvailableOutputFormats`.
  */
-export const contentTypeIsImageForSharp = new RegExp(`^(image\\/)(${Object.keys(Media.AvailableOutputFormats).join('|')})`); // Checks if will be image
+export const contentTypeIsImageForSharp = new RegExp(`^(image\\/)(${Object.keys(AvailableOutputFormats).join('|')})`); // Checks if will be image
 /**
  * Regular expression that tests whether a MIME content-type string represents
  * a JPEG image (either `image/jpeg` or `image/jpg`).
